@@ -9,20 +9,21 @@
     },
     "prettier": "@netblink/nb-prettier",
     "devDependencies": {
-        "@inertiajs/vue3": "^2.0.0",
+        "@inertiajs/vue3": "^2.0",
         "@prettier/plugin-php": "^0.21.0",
         "@shufo/prettier-plugin-blade": "^1.14.1",
-        "@tailwindcss/forms": "^0.5.3",
-        "@vitejs/plugin-vue": "^4.0.0",
-        "autoprefixer": "^10.4.12",
+        "@tailwindcss/forms": "^0.5.7",
+        "@tailwindcss/typography": "^0.5.10",
+        "@vitejs/plugin-vue": "^5.0.0",
+        "autoprefixer": "^10.4.16",
         "axios": "^1.1.2",
         "laravel-vite-plugin": "^0.7.2",
-        "postcss": "^8.4.18",
+        "postcss": "^8.4.32",
         "prettier": "^3.2.5",
         "prettier-plugin-tailwindcss": "^0.5.11",
-        "tailwindcss": "^3.2.1",
+        "tailwindcss": "^3.4.0",
         "vite": "^4.0.0",
-        "vue": "^3.2.41"
+        "vue": "^3.3.13"
     },
     "dependencies": {
         "@fortawesome/fontawesome-svg-core": "^6.4.0",
@@ -35,6 +36,8 @@
         "@fullcalendar/resource": "^6.1.13",
         "@fullcalendar/resource-timeline": "^6.1.13",
         "@fullcalendar/vue3": "^6.1.13",
+        "@inertiajs/inertia-vue3": "^0.6.0",
+        "@inertiajs/progress": "^0.2.7",
         "@mysticwave/collect-me": "^1.0.5",
         "@netblink/nb-prettier": "github:NetBlink/nb-prettier",
         "@netblink/vue-components": "^1.0.3",
